pyproject.toml
[project]
name = "financial-rag"
version = "1.0.0"
description = "A RAG system for querying personal financial PDF documents"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Belden Schroeder"}
]
keywords = ["rag", "langchain", "chromadb", "financial", "pdf", "mcp"]

dependencies = [
    # Core RAG dependencies
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-chroma>=0.1.4",
    "langchain-anthropic>=0.3.0",
    "langchain-huggingface>=0.1.0",
    
    # Vector store
    "chromadb>=0.5.0",
    
    # PDF processing
    "pypdf>=4.0.0",
    "pdfplumber>=0.11.0",
    
    # Embeddings (runs locally, free)
    "sentence-transformers>=3.0.0",
    
    # MCP server for Claude Desktop
    "mcp>=1.0.0",
    
    # File watching for auto-sync
    "watchdog>=4.0.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",  # For nice terminal output
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
]

[project.scripts]
# Command-line entry points
financial-rag = "financial_rag.rag_pipeline:main"
financial-watcher = "financial_rag.file_watcher:main"
financial-mcp = "financial_rag.mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/financial_rag"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
